Lift the no-overlap walk out of equals, so another rule can take it (#900)
The walk that certifies two domains are disjoint over runs rather than
values is not about equals: what it needs from its caller is only that
something makes the two operands equal, and a reification condition is
one such thing among others. Element's index-support rule has another --
an index tuple's guard -- and states the same fact per value today.

A pure move into gcs/constraints/innards/, with the doc comment saying
"where they are equal" in place of "under cond", and what each move owes
a caller written out as a list rather than left to equals' emitter to
imply.
